fix: rename directory pages to view
